+++ src/sessionstorer/test/initialize.php
@@ -5,6 +5,9 @@ define("SESSION_STORER_DEFAULT_SESSION_NAME","session");
 define("SESSION_STORER_COOKIE_NAME_SESSION","%session_name%");
 define("SESSION_STORER_COOKIE_NAME_CHECK","check");
 define("SESSION_STORER_SESSION_MAX_LIFETIME",60 * 60 * 24 * 1);  // a day
+
+define("PACKER_USE_COMPRESS",false);
+
 define("CURRENT_TIME",time());
 require("../../stringbuffer/load.php");
 require("../../http/load.php");
@@ -13,8 +16,9 @@ require("../../translate/load.php");
 require("../load.php");
 
 // UF! we need dbmole
-require(dirname(__FILE__)."/../../dbmole/load.php");
-require(dirname(__FILE__)."/../../dbmole/test/connections_and_handler.php");
+require(__DIR__."/../../dbmole/load.php");
+require(__DIR__."/../../dbmole/test/connections_and_handler.php");
+require(__DIR__."/../../packer/load.php");
 
 // rectreating database structures
 $dbmole = PgMole::GetInstance();